// src/assets/types/constants.auth.tsx
export enum constants {
    SIGNUP_START = "SIGNUP_START",
    SIGNUP_SUCCESS = "SIGNUP_SUCCESS",
    SIGNUP_ERROR = "SIGNUP_FAIL",
    SIGNIN_START = "SIGNIN_START",
    SIGNIN_SUCCESS = "SIGNIN_SUCCESS",
    SIGNIN_ERROR = "SIGNIN_FAIL",
    LOGOUT = "LOGOUT",
    REFRESH_TOKEN = "REFRESH_TOKEN",
}